*** empty log message ***
svn path=/trunk/boinc/; revision=12220

+++ lib/gui_rpc_client.C
@@ -218,7 +218,7 @@ int RPC_CLIENT::authorize(const char* passwd) {
     }
 
     n = snprintf(buf, sizeof(buf), "%s%s", nonce, passwd);
-    if (n >= sizeof(buf)) return ERR_AUTHENTICATOR;
+    if (n >= (int)sizeof(buf)) return ERR_AUTHENTICATOR;
     md5_block((const unsigned char*)buf, (int)strlen(buf), nonce_hash);
     sprintf(buf, "<auth2/>\n<nonce_hash>%s</nonce_hash>\n", nonce_hash);
     retval = rpc.do_rpc(buf);
